Massive Initial Commit
* Changed from SDL2 to SDL3
* Fixed all broken API issues.
* Adjusted Imgui backend to use SDL3
* Changed from SDL2_mixer to SDL3_mixer.
* SDL3_mixer is a completely different paradigm, so all sounds logic had to be replaced with the new style.
* Removed Music and Sound Player and replaced with the AudioPlayer.
* The AudioPlayer follows the new SDL3_mixer tracks and MIX_Audio style.
* Adjusted the Asteroids game to use the new AudioPlayer.

+++ SCION_WINDOW/include/Windowing/Inputs/Mouse.h
@@ -18,7 +18,7 @@ class Mouse
 	const bool IsBtnJustPressed( int btn ) const;
 	const bool IsBtnJustReleased( int btn ) const;
 
-	const std::tuple<int, int> GetMouseScreenPosition();
+	const std::tuple<float, float> GetMouseScreenPosition();
 
 	inline void SetMouseWheelX( int wheel ) { m_WheelX = wheel; }
 	inline void SetMouseWheelY( int wheel ) { m_WheelY = wheel; }
@@ -35,8 +35,8 @@ class Mouse
 		{ SCION_MOUSE_RIGHT, Button{} },
 	};
 
-	int m_X{ 0 };
-	int m_Y{ 0 };
+	float m_X{ 0 };
+	float m_Y{ 0 };
 	int m_WheelX{ 0 };
 	int m_WheelY{ 0 };
 	bool m_bMouseMoving{ false };
